rename to numeric control
